refactor: modularize startup routines
